requirements.txt
dnspython>=2.4.0
requests>=2.31.0
httpx>=0.25.0
python-nmap>=0.7.1
colorama>=0.4.6